inventory/requirements.txt
fastapi
redis-om
uvicorn
requests
python-dotenv
pydantic